Use @Before annotation to inject the system property for testing purposes
